# openapi.yaml
openapi: 3.0.3
info:
  title: Mock Interview Buddy Platform API
  description: >
    REST and WebSocket specification expected by the Mock Interview Buddy frontend application.
    Powers real-time collaborative system design whiteboard interviews between interviewers,
    candidates, observers, and administrators.
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development backend server
  - url: https://api.mockinterviewbuddy.com
    description: Production backend cluster
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/auth/me:
    get:
      tags:
        - Auth & Users
      summary: Get current authenticated user
      description: >
        Returns the profile and role details of the currently authenticated user based on the provided JWT bearer token.
        Expected by `PlatformApi.getCurrentUser()`.
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Authenticated user details
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /api/v1/auth/login:
    post:
      tags:
        - Auth & Users
      summary: User login
      description: Authenticates an interviewer or platform user with email and password to issue a JWT bearer token.
      operationId: login
      security: [] # Public endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "ada@interviews.dev"
                password:
                  type: string
                  format: password
                  example: "password123"
      responses:
        "200":
          description: Successful authentication
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: object
                    required:
                      - user
                      - accessToken
                    properties:
                      user:
                        $ref: "#/components/schemas/User"
                      accessToken:
                        type: string
                        example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
  /api/v1/sessions/{sessionId}/nodes:
    post:
      tags:
        - Canvas Nodes
      summary: Create canvas node
      description: >
        Adds a new system component node to the collaborative canvas.
        Expected by `PlatformApi.createNode(sessionId, node)`.
      operationId: createNode
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCanvasNodeInput"
      responses:
        "201":
          description: Canvas node created
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/CanvasNode"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /api/v1/sessions/{sessionId}/nodes/{nodeId}:
    patch:
      tags:
        - Canvas Nodes
      summary: Update canvas node
      description: >
        Applies a partial update (coordinates, label, notes, dimensions, or kind) to an existing canvas node.
        Expected by `PlatformApi.updateNode(sessionId, nodeId, patch)`.
      operationId: updateNode
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/NodeId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCanvasNodeInput"
      responses:
        "200":
          description: Updated canvas node
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/CanvasNode"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"

    delete:
      tags:
        - Canvas Nodes
      summary: Delete canvas node
      description: >
        Deletes a canvas node and automatically cascades removal to any connected edges.
        Expected by `PlatformApi.deleteNode(sessionId, nodeId)`.
      operationId: deleteNode
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/NodeId"
      responses:
        "204":
          description: Node deleted successfully
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"

  # ============================================================================
  # CANVAS EDGES
  # ============================================================================
  /api/v1/sessions/{sessionId}/edges:
    post:
      tags:
        - Canvas Edges
      summary: Create canvas edge
      description: >
        Connects two nodes with a directional protocol edge. If an identical edge already exists, returns existing.
        Expected by `PlatformApi.createEdge(sessionId, edge)`.
      operationId: createEdge
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCanvasEdgeInput"
      responses:
        "201":
          description: Canvas edge created
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/CanvasEdge"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /api/v1/sessions/{sessionId}/edges/{edgeId}:
    delete:
      tags:
        - Canvas Edges
      summary: Delete canvas edge
      description: >
        Removes a directed edge between components on the canvas.
        Expected by `PlatformApi.deleteEdge(sessionId, edgeId)`.
      operationId: deleteEdge
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/EdgeId"
      responses:
        "204":
          description: Edge deleted successfully
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"

  # ============================================================================
  # CHAT
  # ============================================================================
  /api/v1/sessions/{sessionId}/messages:
    get:
      tags:
        - Chat
      summary: List session chat messages
      description: Retrieves chronological chat message history for the session.
      operationId: listMessages
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of chat messages
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ChatMessage"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"

    post:
      tags:
        - Chat
      summary: Send chat message
      description: >
        Posts a text message into the session chat stream.
        Expected by `PlatformApi.sendMessage(sessionId, body)`.
      operationId: sendMessage
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - body
              properties:
                body:
                  type: string
                  description: Text content of chat message
                  example: "Starting with read-heavy assumptions — 100:1 read to write ratio."
      responses:
        "201":
          description: Chat message sent
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/ChatMessage"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"

  # ============================================================================
  # AUDIT TRAIL
  # ============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter valid JWT token issued on login or session invite join.
  parameters:
    SessionId:
      name: sessionId
      in: path
      required: true
      description: Unique ID of the interview session (e.g. ses_demo)
      schema:
        type: string
        example: "ses_demo"
    NodeId:
      name: nodeId
      in: path
      required: true
      description: Unique ID of the canvas node (e.g. nd_abc123)
      schema:
        type: string
        example: "nd_abc123"
    EdgeId:
      name: edgeId
      in: path
      required: true
      description: Unique ID of the canvas edge (e.g. eg_xyz789)
      schema:
        type: string
        example: "eg_xyz789"
...
  schemas:
...
    Role:
      type: string
      enum:
        - interviewer
        - candidate
        - observer
        - admin
      description: User permission tier in the interview platform.
      example: "interviewer"
...
    NodeKind:
      type: string
      enum:
        - client
        - gateway
        - service
        - database
        - cache
        - queue
        - storage
        - cdn
        - worker
        - external
      description: Catalog architectural archetype for a canvas node.
      example: "database"
    EdgeProtocol:
      type: string
      enum:
        - http
        - grpc
        - async
        - sql
        - stream
      description: Communication protocol representing connection semantics.
      example: "http"

    # --------------------------------------------------------------------------
    # Domain Models
    # --------------------------------------------------------------------------
    User:
      type: object
      required:
        - id
        - displayName
        - email
        - role
        - color
      properties:
        id:
          type: string
          description: Unique user ID
          example: "u_interviewer"
        displayName:
          type: string
          description: User full or display name
          example: "Ada Mensah"
        email:
          type: string
          format: email
          description: User contact email
          example: "ada@interviews.dev"
        role:
          $ref: "#/components/schemas/Role"
        color:
          type: string
          description: Hex color code assigned for cursor, presence, and annotations
          example: "#f0b429"
...
    CanvasNode:
      type: object
      required:
        - id
        - sessionId
        - kind
        - label
        - notes
        - x
        - y
        - width
        - height
        - version
      properties:
        id:
          type: string
          description: Unique canvas node ID
          example: "nd_gateway1"
        sessionId:
          type: string
          description: Session ID containing this node
          example: "ses_demo"
        kind:
          $ref: "#/components/schemas/NodeKind"
        label:
          type: string
          description: Display name of the architecture component
          example: "API Gateway"
        notes:
          type: string
          description: Notes or configuration details for this component
          example: "Rate limiting and SSL termination"
        x:
          type: number
          description: X coordinate on infinite canvas
          example: 330
        y:
          type: number
          description: Y coordinate on infinite canvas
          example: 120
        width:
          type: number
          description: Rendered width in pixels
          example: 176
        height:
          type: number
          description: Rendered height in pixels
          example: 84
        version:
          type: integer
          description: Monotonically increasing version counter for optimistic concurrency
          example: 1
    CanvasEdge:
      type: object
      required:
        - id
        - sessionId
        - fromNodeId
        - toNodeId
        - label
        - protocol
      properties:
        id:
          type: string
          description: Unique edge ID
          example: "eg_1"
        sessionId:
          type: string
          description: Session ID
          example: "ses_demo"
        fromNodeId:
          type: string
          description: Source canvas node ID
          example: "nd_client"
        toNodeId:
          type: string
          description: Target canvas node ID
          example: "nd_gateway"
        label:
          type: string
          description: Optional protocol or interaction note displayed along the edge
          example: "HTTPS"
        protocol:
          $ref: "#/components/schemas/EdgeProtocol"
    ChatMessage:
      type: object
      required:
        - id
        - sessionId
        - authorId
        - authorName
        - body
        - createdAt
      properties:
        id:
          type: string
          description: Unique chat message ID
          example: "msg_123"
        sessionId:
          type: string
          description: Session ID
          example: "ses_demo"
        authorId:
          type: string
          description: User ID who posted the message
          example: "u_candidate"
        authorName:
          type: string
          description: Author display name
          example: "Noah Chen"
        body:
          type: string
          description: Message text
          example: "Starting with read-heavy assumptions — 100:1 read to write ratio."
        createdAt:
          type: string
          format: date-time
          example: "2026-09-15T12:05:00.000Z"
...
    CreateCanvasNodeInput:
      type: object
      required:
        - kind
        - label
        - x
        - y
        - width
        - height
      properties:
        kind:
          $ref: "#/components/schemas/NodeKind"
        label:
          type: string
          example: "Web Client"
        notes:
          type: string
          default: ""
          example: ""
        x:
          type: number
          example: 80
        y:
          type: number
          example: 120
        width:
          type: number
          default: 176
          example: 176
        height:
          type: number
          default: 84
          example: 84
    UpdateCanvasNodeInput:
      type: object
      properties:
        kind:
          $ref: "#/components/schemas/NodeKind"
        label:
          type: string
          example: "API Gateway"
        notes:
          type: string
          example: "Rate limiting enabled"
        x:
          type: number
          example: 350
        y:
          type: number
          example: 140
        width:
          type: number
          example: 176
        height:
          type: number
          example: 84
    CreateCanvasEdgeInput:
      type: object
      required:
        - fromNodeId
        - toNodeId
        - protocol
      properties:
        fromNodeId:
          type: string
          description: ID of the source canvas node
          example: "nd_client"
        toNodeId:
          type: string
          description: ID of the destination canvas node
          example: "nd_gateway"
        label:
          type: string
          default: ""
          description: Edge label (e.g. HTTPS, REST, SQL)
          example: "HTTPS"
        protocol:
          $ref: "#/components/schemas/EdgeProtocol"
...
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: "BAD_REQUEST"
        message:
          type: string
          example: "Invalid input parameters"
        details:
          type: object
          nullable: true
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"

  # ============================================================================
  # STANDARD RESPONSES
  # ============================================================================
  responses:
    BadRequestError:
      description: Invalid request payload, missing fields, or validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    UnauthorizedError:
      description: Authentication is missing, invalid, or expired
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    ForbiddenError:
      description: Authenticated user lacks required permissions for this operation/session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFoundError:
      description: The requested resource (session, node, edge, or snapshot) was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"